Add Data.DoubleMap module.
