+++ src/Data/Delegation.hs
@@ -29,6 +29,7 @@ import Data.SafeCopy (base, deriveSafeCopy)
 
 import Control.Monad.Reader
 
+import Data.DoubleMap
 import Types
 
 
@@ -82,19 +83,6 @@ data Delegations = Delegations {
 
 -- * helpers
 
-lookupDoubleMap :: (Ord k1, Ord k2) => k1 -> k2 -> Map k1 (Map k2 v) -> Maybe v
-lookupDoubleMap k1 k2 m = Map.lookup k1 m >>= Map.lookup k2
-
-insertDoubleMap :: (Ord k1, Ord k2) => k1 -> k2 -> v -> Map k1 (Map k2 v) -> Map k1 (Map k2 v)
-insertDoubleMap k1 k2 v m = case Map.lookup k1 m of
-    Nothing -> Map.insert k1 (Map.singleton k2 v) m
-    Just m' -> Map.insert k1 (Map.insert k2 v m') m
-
-deleteDoubleMap :: (Ord k1, Ord k2) => k1 -> k2 -> Map k1 (Map k2 v) -> Map k1 (Map k2 v)
-deleteDoubleMap k1 k2 m = m
-                        & at k1 . _Just . at k2 .~ Nothing
-                        & at k1 %~ deleteEmpty
-
 lookupDMap :: Delegatee U -> S -> DelegationMap -> Maybe (Delegate U)
 lookupDMap d s (DelegationMap dmap) = lookupDoubleMap d s dmap
 
@@ -121,10 +109,6 @@ setDelegation f dscope t (Delegations (DelegationMap dmap) (CoDelegationMap coDm
                                            & at to' . _Just . at dscope %~ deleteEmpty
                                            & at to' %~ deleteEmpty) mOldTo
     coDmap1 = insertDoubleMap to dscope (Set.insert from $ fromMaybe Set.empty (lookupDoubleMap to dscope coDmap0)) coDmap0
-
-deleteEmpty :: Foldable t => Maybe (t a) -> Maybe (t a)
-deleteEmpty (Just s) | null s = Nothing
-deleteEmpty x                 = x
 
 -- | FIXME: if we mention the delegate here, not just the delegatee, we can confirm that is is the
 -- one we expect, and we will catch errors more local to their source.
