+++ tests/AllModulesSpec.hs
@@ -17,6 +17,7 @@ import Config ()
 import Daemon ()
 import Data.Avatar ()
 import Data.Delegation ()
+import Data.DoubleMap ()
 import Data.Markdown ()
 import Data.UriPath ()
 import DemoData ()
